+++ pkg/lint/rules/avoid_logical_constructs_test.go
@@ -1,0 +1,42 @@
+package rules
+
+import (
+	"testing"
+
+	"github.com/giantswarm/schemalint/pkg/lint"
+)
+
+func TestAvoidLogicalContructs(t *testing.T) {
+	testCases := []struct {
+		name        string
+		schemaPath  string
+		nViolations int
+	}{
+		{
+			name:        "case 0: uses if, then, else",
+			schemaPath:  "testdata/avoid_logical_constructs/incorrect.json",
+			nViolations: 1,
+		},
+		{
+			name:        "case 1: does not use if, then, else",
+			schemaPath:  "testdata/avoid_logical_constructs/correct.json",
+			nViolations: 0,
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			schema, err := lint.Compile(tc.schemaPath)
+			if err != nil {
+				t.Fatalf("Unexpected parsing error in test case '%s': %s", tc.name, err)
+			}
+
+			rule := AvoidLogicalConstruct{}
+			ruleResults := rule.Verify(schema)
+
+			if len(ruleResults.Violations) != tc.nViolations {
+				t.Fatalf("Unexpected number of rule violations in test case '%s': Expected %d, got %d", tc.name, tc.nViolations, len(ruleResults.Violations))
+			}
+		})
+	}
+}